Add notes and a small cleanup

+++ parser/src/parser.rs
@@ -346,8 +346,8 @@ where
         let token = self.lookahead.take();
 
         if let Some(token) = token {
-            let text = &self.input[token.start.absolute as usize
-                ..token.start.absolute as usize + token.value.len as usize];
+            let text = &self.input
+                [token.start.absolute as usize..(token.start.absolute + token.value.len) as usize];
             self.builder.token(token.value.kind.into(), text.into());
             self.past_tokens.push_front(token);
             self.lookahead = self.iter.next()
